requirements.txt
appdirs==1.4.3
attrs==19.3.0
Babel==2.7.0
black==19.10b0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
codecov==2.0.15
coverage==4.5.4
dj-database-url==0.5.0
Django==2.2.10
django-crispy-forms==1.8.1
django-email-auth==0.1.5
django-heroku==0.3.1
django-phonenumber-field==3.0.1
entrypoints==0.3
flake8==3.7.9
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.1.0
itypes==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
mccabe==0.6.1
model-mommy==2.0.0
more-itertools==8.0.0
openapi-codec==1.3.2
packaging==19.2
pathspec==0.6.0
phonenumbers==8.11.0
pluggy==0.13.1
psycopg2==2.7.5
psycopg2-binary==2.8.3
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.5
pytest==5.3.1
pytest-django==3.7.0
python-dateutil==2.8.1
pytz==2019.3
regex==2019.11.1
requests==2.22.0
simplejson==3.17.0
six==1.13.0
sqlparse==0.3.0
text-unidecode==1.3
toml==0.10.0
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
whitenoise==4.1.4
zipp==0.6.0
